Refactor:added all dependencies and added all files and directory system
