+++ Readme.md
@@ -1,0 +1,73 @@
+
+--Architecture--dummy--
+com.example.zip-feast
+├── data
+│   ├── model
+│   │   ├── Product.kt
+│   │   ├── User.kt
+│   │   ├── Order.kt
+│   │   └── (Other data classes)
+│   ├── repository
+│   │   ├── ProductRepository.kt
+│   │   ├── UserRepository.kt
+│   │   ├── OrderRepository.kt
+│   │   └── (Other repositories)
+│   ├── network
+│   │   ├── ApiService.kt
+│   │   ├── NetworkModule.kt
+│   │   └── (Retrofit or other network-related classes)
+│   ├── local
+│   │   ├── dao
+│   │   │   ├── ProductDao.kt
+│   │   │   ├── UserDao.kt
+│   │   │   ├── OrderDao.kt
+│   │   │   └── (Other DAO interfaces)
+│   │   └── AppDatabase.kt
+│   │   └── (Room database setup)
+│   └── remote
+│       ├── RemoteDataSource.kt
+│       └── (Other remote data source related classes)
+├── ui
+│   ├── product
+│   │   ├── ProductFragment.kt
+│   │   ├── ProductViewModel.kt
+│   │   ├── ProductAdapter.kt
+│   │   └── (Other product-related UI classes)
+│   ├── cart
+│   │   ├── CartFragment.kt
+│   │   ├── CartViewModel.kt
+│   │   ├── CartAdapter.kt
+│   │   └── (Other cart-related UI classes)
+│   ├── order
+│   │   ├── OrderFragment.kt
+│   │   ├── OrderViewModel.kt
+│   │   ├── OrderAdapter.kt
+│   │   └── (Other order-related UI classes)
+│   ├── profile
+│   │   ├── ProfileFragment.kt
+│   │   ├── ProfileViewModel.kt
+│   │   └── (Other profile-related UI classes)
+│   ├── auth
+│   │   ├── LoginFragment.kt
+│   │   ├── RegisterFragment.kt
+│   │   ├── AuthViewModel.kt
+│   │   └── (Other authentication-related UI classes)
+│   └── common
+│       ├── BaseFragment.kt
+│       ├── BaseActivity.kt
+│       └── (Common UI components)
+├── viewmodel
+│   └── (Any shared ViewModel classes or base ViewModel classes)
+├── di
+│   ├── AppModule.kt
+│   ├── RepositoryModule.kt
+│   ├── NetworkModule.kt
+│   ├── ViewModelModule.kt
+│   └── (Dependency injection setup with Dagger/Hilt)
+├── utils
+│   ├── Constants.kt
+│   ├── Extensions.kt
+│   └── (Other utility classes)
+└── application
+└── EcommerceApplication.kt
+└── (Custom application class for initialization)
