+++ app/src/main/java/com/zip_feast/application/ZipFeastActivity.kt
@@ -1,14 +1,13 @@
-package com.zip_feast
+package com.zip_feast.application
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import com.zip_feast.screens.LoginScreen
-import com.zip_feast.screens.RegistrationScreen
+import com.zip_feast.ui.auth.screens.RegistrationScreen
 import com.zip_feast.ui.theme.ZipFeastTheme
 
-class MainActivity : ComponentActivity() {
+class ZipFeastActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
