+++ app/src/main/java/com/zip_feast/ui/auth/screens/LoginScreen.kt
@@ -1,4 +1,4 @@
-package com.zip_feast.screens
+package com.zip_feast.ui.auth.screens
 
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.foundation.layout.Arrangement
